generalize WhitelistedOutputTypeCompletionProposalPredicate
